pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["netbox_pathways", "netbox_pathways.*"]

[tool.setuptools.package-data]
netbox_pathways = [
    "templates/**/*",
    "static/**/dist/*",
    "static/**/vendor/*",
    "static/**/css/*",
    "static/**/qgis/*",
]

[project]
name = "netbox-pathways"
version = "0.1.0"
description = "NetBox plugin for physical cable plant infrastructure documentation with GIS capabilities"
readme = "README.md"
authors = [
    { name = "Jonathan Senecal", email = "contact@jonathansenecal.com" },
]
license-files = ["LICENSE"]
requires-python = ">=3.12"

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

dependencies = [
    "djangorestframework-gis>=1.2.0",
    "networkx>=3.0",
]

[project.optional-dependencies]
dev = [
    "bumpver",
    "pre-commit>=4.0.0",
    "pytest",
    "pytest-django>=4.5.0",
    "pytest-cov>=3.0.0",
    "ruff",
]
docs = [
    "zensical",
]

[project.urls]
Homepage = "https://github.com/jsenecal/netbox-pathways"
Source = "https://github.com/jsenecal/netbox-pathways"
Tracker = "https://github.com/jsenecal/netbox-pathways/issues"
Documentation = "https://jsenecal.github.io/netbox-pathways/"

[tool.bumpver]
current_version = "0.1.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "chore: bump version {old_version} -> {new_version}"
tag_pattern = "vMAJOR.MINOR.PATCH"
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
    'version = "{version}"',
]
"netbox_pathways/__init__.py" = [
    '__version__ = "{version}"',
]
"CHANGELOG.md" = [
    '## \[Unreleased\]',
    '## \[{version}\] - {now:%Y-%m-%d}',
]

[tool.ruff]
line-length = 120
target-version = "py312"
extend-exclude = ["**/migrations/*.py"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "DJ", "PIE"]
ignore = ["E501", "N806", "S101", "DJ01"]

[tool.ruff.lint.per-file-ignores]
"migrations/*" = ["N806", "N999"]
"tests/*" = ["S101", "S105", "S106", "E402", "F841"]
"netbox_pathways/management/commands/generate_sample_data.py" = ["S311", "B007"]
"netbox_pathways/management/commands/import_geodata.py" = ["S603", "S607"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "netbox.settings"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=netbox_pathways --cov-report=term-missing --reuse-db"
